converted to pyproject.toml added old benchmarks
